backend/requirements.txt
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
click==7.1.2
coverage==5.5
Django==3.1.6
django-utils==0.0.2
dnspython==1.16.0
gunicorn==20.0.4
Flask==1.1.2
Flask-Cors==3.0.10
iniconfig==1.1.1
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.8
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pylint==2.6.0
pymongo==3.11.2
pyparsing==2.4.7
pytest==6.2.1
python-dotenv==0.17.0
pytz==2021.1
random2==1.0.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
